Use existing values as defaults during prepare
Prepare is only called when a config file is used, and only after parm
parsing is completed. There may be values already set during parm
parsing, so it's necessary to load them before assigning the values
passed into prepare. This makes it possible to only have some values
populated in the config file, while the rest are merged during prepare.